LITE-20908: update the entrypoint.sh and Dockerfile to manually git checkout

+++ entrypoint.sh
@@ -3,27 +3,52 @@ set -e
 
 EXTENSION_DIR=${EXTENSION_DIR:-'/extension'}
 
-DIST_DIR=$EXTENSION_DIR/dist
+if [[ "$EXTENSION_DIR"  != /* ]]; then
+    echo "Extension directory ($EXTENSION_DIR) must be an absolute path"
+    exit 1
+fi
 
-echo "Installing the extension package from $EXTENSION_DIR..."
+if [[ "$@" == *"cextrun"* ]]; then
 
-pushd .
+    if [ -n "${REPOSITORY_URL+set}" ] && [ -n "${COMMIT_ID+set}" ]; then  
+        git clone $REPOSITORY_URL $EXTENSION_DIR
+        git -C $EXTENSION_DIR checkout $COMMIT_ID
+    fi
 
-cd $EXTENSION_DIR
+    if [ ! -d $EXTENSION_DIR ]; then
+        echo "Extension directory ($EXTENSION_DIR) not found"
+        exit 1
+    fi
 
-if [ -d "$DIST_DIR" ]; then rm -Rf $DIST_DIR; fi
+    if [ ! -f $EXTENSION_DIR/pyproject.toml ]; then
+        echo "pyproject.toml not found in $EXTENSION_DIR"
+        exit 1
+    fi
 
-poetry build
+    DIST_DIR=$EXTENSION_DIR/dist
 
-pip install -U pip && pip install -U --force-reinstall $DIST_DIR/*.whl
+    echo "Installing the extension package from $EXTENSION_DIR..."
+    
+    pushd .
+    
+    cd $EXTENSION_DIR
 
-popd
+    if [ -d "$DIST_DIR" ]; then rm -Rf $DIST_DIR; fi
 
-echo "Extension installed."
+    poetry build
+
+    pip install -U pip && pip install -U --force-reinstall $DIST_DIR/*.whl
+
+    popd
+
+    echo "Extension installed."
+fi
 
 if [[ "$@" == *bash* ]]; then
-echo "In order to run the extension manually run the command cextrun"
-echo "In order to run the extension in debug mode, please use cextrun -d command"
-echo "In the case that you modified dependencies or want to install development dependencies run the command: poetry install"
+    echo "To install the extension execute poetry build inside /extension folder"
+    echo "In order to run the extension manually run the command cextrun"
+    echo "In order to run the extension in debug mode, please use cextrun -d command"
+    echo "In the case that you modified dependencies or want to install development dependencies run the command: poetry install"
 fi
+
 exec "$@"
